ARM64 final profiles and tiny fixes in AMD64
The profile-arm64-minimum.txt now runs ARM64 FreeBSD on QEMU using UFS
storage and EFI booting with no issues. Devices such as pl011 gpio and
acpi are needed for booting and console. Devices such as clk, phy and
regulator are needed by the ARM64 kernel to work.

Added the profile-arm64-jail.txt profile, which is a copy of
amd64-jail.txt, minus AMD64 bits and includes the ARM64 bits.

Fixed an extra space in AMD64 minimum profile.

Added a line for META_MODE fix, which is not clear yet, for why it
happens on ARM64, but in case someone else needs it, it's there.

profile-arm64-{minimum,jail}.txt have been tested on QEMU, all features
(zfs, epair, bridge) work fine.